percentage pruning ffg

+++ inference/utils.py
@@ -90,7 +90,7 @@ class L1UnstructuredFFG(prune.BasePruningMethod):
             else:
                 mask[idx[:,0], mask[:,1]] = 0.
             masks.append(mask.flatten())
-        self.mask = torch.cat(masks)
+        self.mask = torch.cat(masks).type(mu.dtype)
 
     def compute_mask(self, tensor, default_mask):
         import pdb; pdb.set_trace()
